requirements.txt
pytest
fastapi
pydantic
uvicorn[standard]
pillow
python-multipart
openai